% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiery.R
\docType{data}
\name{app}
\alias{app}
\title{The main app}
\format{An object of class \code{Fire} (inherits from \code{R6}) of length 43.}
\usage{
app
}
\description{
This is a fiery instance that can be run with \code{ignite()}. See \link[fiery:Fire]{fiery::Fire}
}
\keyword{datasets}
